+++ bin/heta-build.js
@@ -28,7 +28,7 @@ program
   .option('-S, --skip-export', 'do not export files to local directory')
   .option('-L, --log-mode <never|error|always>', 'When to create log file.')
   .option('-d, --debug', 'If set the raw module output will be stored in "meta".')
-  .option('--ss-only', 'Run in SimSolver supporting mode: skip declared exports, add default export to SimSolver.')
+  .option('--julia-only', 'Run in Julia supporting mode: skip declared exports.')
   .option('--dist-dir <filepath>', 'Set export directory path, where to store distributives.')
   .option('--meta-dir <filepath>', 'Set meta directory path.')
   // moduleImport